requirements.txt
# Install into a venv (see comments at top of extract_exercises.py).
pymupdf>=1.24.0
Pillow>=10.0.0
reportlab>=4.0.0
python-dotenv>=1.0.0
openai>=1.40.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
jinja2>=3.1.0
python-multipart>=0.0.9
